+++ crates/pow/src/lib.rs
@@ -1,4 +1,3 @@
-use std::num::NonZeroUsize;
 use std::sync::{Arc, Mutex};
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
@@ -37,11 +36,9 @@ type Domain = [u8; 32];
 pub struct PoWRateLimiterConfig {
     /// The lifetime for challenges. After this time, challenges are considered expired.
     pub challenge_lifetime: Duration,
-    /// Determines how much the difficulty increases with the amount of active challenges. The
-    /// difficulty is computed as `num_active_challenges << growth_rate`.
-    pub growth_rate: NonZeroUsize,
-    /// Sets the `max_target` used for challenges. The initial target (with difficulty = 1) for
-    /// challenges will be `u64::MAX >> baseline`.
+    /// Determines how much the difficulty increases with the number of active challenges.
+    pub growth_rate: f64,
+    /// Sets the baseline difficulty bits when there are no active challenges.
     pub baseline: u8,
     /// The interval at which the challenge cache is cleaned up. Only expired challenges are
     /// removed during cleanup.
@@ -66,11 +63,21 @@ impl PoWRateLimiter {
         }
     }
 
-    /// Generates a new challenge.
+    /// Generates a new challenge with a difficulty that will depend on the number of active
+    /// challenges for the given domain and the request complexity.
+    ///
+    /// # Arguments
+    /// * `requestor` - A unique identifier for the user that is requesting the challenge.
+    /// * `domain` - A unique identifier for the service that is requesting the challenge.
+    /// * `request_complexity` - A measure of the complexity of the request. Must be greater than 0.
+    ///
+    /// # Panics
+    /// Panics if the request complexity is 0.
     pub fn build_challenge(
         &self,
         requestor: impl Into<Requestor>,
         domain: impl Into<Domain>,
+        request_complexity: u64,
     ) -> Challenge {
         let current_time = SystemTime::now()
             .duration_since(UNIX_EPOCH)
@@ -78,30 +85,39 @@ impl PoWRateLimiter {
             .as_secs();
         let requestor = requestor.into();
         let domain = domain.into();
-        let target = self.get_challenge_target(&domain);
+        let target = self.get_challenge_target(&domain, request_complexity);
 
-        Challenge::new(target, current_time, requestor, domain, self.secret)
+        Challenge::new(target, current_time, request_complexity, requestor, domain, self.secret)
     }
 
     /// Computes the target for a given domain by checking the amount of active challenges in the
-    /// cache. This sets the difficulty of the challenge.
+    /// cache and the given request complexity.
     ///
-    /// It is computed as:
-    /// `max_target / difficulty`
+    /// The target is computed as: `target = u64::MAX / request_difficulty`
     ///
     /// Where:
-    /// * `max_target = u64::MAX >> baseline`
-    /// * `difficulty = max(num_active_challenges << growth_rate, 1)`
-    fn get_challenge_target(&self, domain: &Domain) -> u64 {
+    /// * `request_difficulty = load_difficulty * request_complexity`
+    /// * `load_difficulty = 2^baseline * ceil((num_active_challenges + 1) * growth_rate)`
+    fn get_challenge_target(&self, domain: &Domain, request_complexity: u64) -> u64 {
+        if request_complexity == 0 {
+            return u64::MAX;
+        }
+
         let num_challenges = self
             .challenges
             .lock()
             .expect("challenge cache lock should not be poisoned")
             .num_challenges_for_domain(domain);
 
-        let max_target = u64::MAX >> self.config.baseline;
-        let difficulty = usize::max(num_challenges << self.config.growth_rate.get(), 1);
-        max_target / difficulty as u64
+        #[allow(clippy::cast_precision_loss, reason = "num_challenges is smaller than f64::MAX")]
+        #[allow(clippy::cast_sign_loss, reason = "growth_rate and num_challenges are positive")]
+        let growth_multiplier =
+            ((num_challenges + 1) as f64 * self.config.growth_rate).ceil() as u64;
+        let load_difficulty =
+            2_u64.pow(self.config.baseline.into()).saturating_mul(growth_multiplier);
+        let request_difficulty = load_difficulty.saturating_mul(request_complexity);
+
+        u64::MAX / request_difficulty
     }
 
     /// Submits a challenge.
@@ -124,6 +140,7 @@ impl PoWRateLimiter {
         challenge: &Challenge,
         nonce: u64,
         current_time: u64,
+        request_complexity: u64,
     ) -> Result<(), ChallengeError> {
         challenge.verify_signature(self.secret)?;
 
@@ -139,7 +156,8 @@ impl PoWRateLimiter {
         let valid_requestor = requestor == challenge.requestor;
         let valid_domain = domain == challenge.domain;
         let valid_nonce = challenge.validate_pow(nonce);
-        if !(valid_nonce && valid_requestor && valid_domain) {
+        let valid_request_complexity = challenge.request_complexity == request_complexity;
+        if !(valid_nonce && valid_requestor && valid_domain && valid_request_complexity) {
             return Err(ChallengeError::InvalidPoW);
         }
 
@@ -200,8 +218,8 @@ mod tests {
             secret,
             PoWRateLimiterConfig {
                 challenge_lifetime: Duration::from_secs(30),
+                growth_rate: 1.0,
                 cleanup_interval: Duration::from_millis(500),
-                growth_rate: NonZeroUsize::new(2).unwrap(),
                 baseline: 0,
             },
         )
@@ -212,18 +230,32 @@ mod tests {
         let pow = create_test_pow();
         let domain = [1u8; 32];
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
-
         let requestor = [0u8; 32];
-        let challenge = pow.build_challenge(requestor, domain);
+        let request_complexity = 1;
+        let challenge = pow.build_challenge(requestor, domain, request_complexity);
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         // Submit challenge with correct nonce - should succeed
-        let result = pow.submit_challenge(requestor, domain, &challenge, nonce, current_time);
+        let result = pow.submit_challenge(
+            requestor,
+            domain,
+            &challenge,
+            nonce,
+            current_time,
+            request_complexity,
+        );
         assert!(result.is_ok());
 
         // Try to use the same challenge again with another requestor - should fail
         let requestor = [1u8; 32];
-        let result = pow.submit_challenge(requestor, domain, &challenge, nonce, current_time);
+        let result = pow.submit_challenge(
+            requestor,
+            domain,
+            &challenge,
+            nonce,
+            current_time,
+            request_complexity,
+        );
         assert!(result.is_err());
     }
 
@@ -232,9 +264,10 @@ mod tests {
         let pow = create_test_pow();
         let domain = [1u8; 32];
         let requestor = [0u8; 32];
+        let request_complexity = 1;
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
 
-        let challenge = pow.build_challenge(requestor, domain);
+        let challenge = pow.build_challenge(requestor, domain, request_complexity);
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         // Submit challenge with expired timestamp - should fail
@@ -244,11 +277,19 @@ mod tests {
             &challenge,
             nonce,
             current_time + pow.config.challenge_lifetime.as_secs() + 1,
+            request_complexity,
         );
         assert!(result.is_err());
 
         // Submit challenge with correct timestamp - should succeed
-        let result = pow.submit_challenge(requestor, domain, &challenge, nonce, current_time);
+        let result = pow.submit_challenge(
+            requestor,
+            domain,
+            &challenge,
+            nonce,
+            current_time,
+            request_complexity,
+        );
         assert!(result.is_ok());
     }
 
@@ -257,22 +298,25 @@ mod tests {
         let pow = create_test_pow();
         let domain = [1u8; 32];
         let requestor = [0u8; 32];
+        let request_complexity = 1;
 
         // Solve first challenge
-        let challenge = pow.build_challenge(requestor, domain);
+        let challenge = pow.build_challenge(requestor, domain, request_complexity);
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         let time_1 = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
-        let result = pow.submit_challenge(requestor, domain, &challenge, nonce, time_1);
+        let result =
+            pow.submit_challenge(requestor, domain, &challenge, nonce, time_1, request_complexity);
         assert!(result.is_ok());
 
         // Try to submit second challenge - should fail because of rate limiting
         tokio::time::sleep(pow.config.cleanup_interval).await;
-        let challenge = pow.build_challenge(requestor, domain);
+        let challenge = pow.build_challenge(requestor, domain, request_complexity);
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         let time_2 = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
-        let result = pow.submit_challenge(requestor, domain, &challenge, nonce, time_2);
+        let result =
+            pow.submit_challenge(requestor, domain, &challenge, nonce, time_2, request_complexity);
         assert!(result.is_err());
         let remaining_time = time_1
             + pow.config.challenge_lifetime.as_secs()
@@ -287,30 +331,68 @@ mod tests {
 
         // Try to submit it using a different api key - should succeed
         let domain = [2u8; 32];
-        let challenge = pow.build_challenge(requestor, domain);
+        let challenge = pow.build_challenge(requestor, domain, request_complexity);
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
-        let result = pow.submit_challenge(requestor, domain, &challenge, nonce, time_2);
+        let result =
+            pow.submit_challenge(requestor, domain, &challenge, nonce, time_2, request_complexity);
         assert!(result.is_ok());
     }
 
     #[tokio::test]
     async fn submit_challenge_and_check_difficulty() {
-        let mut pow = create_test_pow();
-        pow.config.growth_rate = NonZeroUsize::new(1).unwrap();
+        let pow = create_test_pow();
         let domain = [1u8; 32];
         let requestor = [0u8; 32];
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
+        let request_complexity = 1;
 
-        assert_eq!(pow.get_challenge_target(&domain), u64::MAX >> pow.config.baseline);
+        assert_eq!(
+            pow.get_challenge_target(&domain, request_complexity),
+            u64::MAX >> pow.config.baseline
+        );
 
-        let challenge = pow.build_challenge(requestor, domain);
+        let challenge = pow.build_challenge(requestor, domain, request_complexity);
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
-        pow.submit_challenge(requestor, domain, &challenge, nonce, current_time)
-            .unwrap();
+        pow.submit_challenge(
+            requestor,
+            domain,
+            &challenge,
+            nonce,
+            current_time,
+            request_complexity,
+        )
+        .unwrap();
 
         assert_eq!(pow.challenges.lock().unwrap().num_challenges_for_domain(&domain), 1);
-        assert_eq!(pow.get_challenge_target(&domain), (u64::MAX >> pow.config.baseline) / 2);
+        assert_eq!(
+            pow.get_challenge_target(&domain, request_complexity),
+            (u64::MAX >> pow.config.baseline) / 2
+        );
+    }
+
+    #[tokio::test]
+    async fn difficulty_increases_with_request_complexity() {
+        let pow = create_test_pow();
+        let domain = [1u8; 32];
+
+        // test: request complexity 1 should have difficulty 1
+        let request_complexity = 1;
+
+        let difficulty = 1;
+        assert_eq!(
+            pow.get_challenge_target(&domain, request_complexity),
+            (u64::MAX >> pow.config.baseline) / difficulty
+        );
+
+        // test: request complexity 3 should have difficulty 3
+        let request_complexity = 3;
+
+        let difficulty = 3;
+        assert_eq!(
+            pow.get_challenge_target(&domain, request_complexity),
+            (u64::MAX >> pow.config.baseline) / difficulty
+        );
     }
 
     #[tokio::test]
@@ -318,18 +400,39 @@ mod tests {
         let pow = create_test_pow();
         let domain = [1u8; 32];
         let requestor = [0u8; 32];
+        let request_complexity = 1;
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
         let target = u64::MAX;
 
         // build challenge manually with past timestamp to ensure that expires in 1 second
         let timestamp = current_time - pow.config.challenge_lifetime.as_secs();
-        let signature =
-            Challenge::compute_signature(pow.secret, target, timestamp, &requestor, &domain);
-        let challenge = Challenge::from_parts(target, timestamp, requestor, domain, signature);
+        let signature = Challenge::compute_signature(
+            pow.secret,
+            target,
+            timestamp,
+            request_complexity,
+            &requestor,
+            &domain,
+        );
+        let challenge = Challenge::from_parts(
+            target,
+            timestamp,
+            request_complexity,
+            requestor,
+            domain,
+            signature,
+        );
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
-        pow.submit_challenge(requestor, domain, &challenge, nonce, current_time)
-            .unwrap();
+        pow.submit_challenge(
+            requestor,
+            domain,
+            &challenge,
+            nonce,
+            current_time,
+            request_complexity,
+        )
+        .unwrap();
 
         // wait for cleanup
         tokio::time::sleep(pow.config.cleanup_interval + Duration::from_secs(1)).await;
@@ -345,12 +448,19 @@ mod tests {
         assert_eq!(pow.challenges.lock().unwrap().num_challenges_for_domain(&domain), 0);
 
         // submit second challenge - should succeed
-        let challenge = pow.build_challenge(requestor, domain);
+        let challenge = pow.build_challenge(requestor, domain, request_complexity);
         let nonce = find_pow_solution(&challenge, 10000).expect("Should find solution");
 
         let current_time = SystemTime::now().duration_since(UNIX_EPOCH).unwrap().as_secs();
-        pow.submit_challenge(requestor, domain, &challenge, nonce, current_time)
-            .unwrap();
+        pow.submit_challenge(
+            requestor,
+            domain,
+            &challenge,
+            nonce,
+            current_time,
+            request_complexity,
+        )
+        .unwrap();
 
         assert_ne!(
             pow.challenges